Added query() and fields() methods to BaseNodeSource to have access to alias by default. Added Node source plugin as most basic instantiation of BaseNodeSource to allow class-free implementations of migrations that can get by with just a YAML config file. Updated SPPA migrations accordingly.
