+++ docroot/modules/custom/sitenow_migrate/src/Plugin/migrate/source/Node.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Drupal\sitenow_migrate\Plugin\migrate\source;
+
+use Drupal\migrate\Event\MigrateImportEvent;
+use Drupal\migrate\Row;
+
+/**
+ * Basic implementation of the source plugin.
+ *
+ * @MigrateSource(
+ *  id = "sitenow_node",
+ *  source_module = "node"
+ * )
+ */
+class Node extends BaseNodeSource {
+
+}
