+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/config/split/migrate_plus.migration.sppa_consulting_firm.yml
@@ -13,7 +13,7 @@ migration_tags:
 migration_group: sitenow_migrate
 label: Consulting Firms
 source:
-  plugin: sppa_consulting_firm
+  plugin: sitenow_node
   node_type: consult_firm
   constants:
     uid_admin: 1
