+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/config/split/migrate_plus.migration.sppa_consulting_firm_redirects.yml
@@ -10,7 +10,7 @@ migration_tags:
 migration_group: sitenow_migrate
 label: 'Consulting Firm Redirects'
 source:
-  plugin: sppa_consulting_firm
+  plugin: sitenow_node
   node_type: consult_firm
   constants:
     redirect_prefix: 'internal:/node/'
