+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/config/split/migrate_plus.migration.sppa_potential_employer.yml
@@ -13,7 +13,7 @@ migration_tags:
 migration_group: sitenow_migrate
 label: Employers
 source:
-  plugin: sppa_potential_employer
+  plugin: sitenow_node
   node_type: employer
   constants:
     uid_admin: 1
