+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/config/split/migrate_plus.migration.sppa_potential_employer_redirects.yml
@@ -10,7 +10,7 @@ migration_tags:
 migration_group: sitenow_migrate
 label: 'Potential Employer Redirects'
 source:
-  plugin: sppa_potential_employer
+  plugin: sitenow_node
   node_type: employer
   constants:
     redirect_prefix: 'internal:/node/'
