+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/src/Plugin/migrate/source/Articles.php
@@ -21,25 +21,6 @@ class Articles extends BaseNodeSource {
   /**
    * {@inheritdoc}
    */
-  public function query() {
-    $query = parent::query();
-    $query->leftJoin('url_alias', 'alias', "alias.source = CONCAT('node/', n.nid)");
-    $query->fields('alias', ['alias']);
-    return $query;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function fields() {
-    $fields = parent::fields();
-    $fields['alias'] = $this->t('The URL alias for this node.');
-    return $fields;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function prepareRow(Row $row) {
     parent::prepareRow($row);
     $body = $row->getSourceProperty('body');
