--- docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/src/Plugin/migrate/source/PotentialEmployer.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace Drupal\sppa_migrate\Plugin\migrate\source;
-
-use Drupal\sitenow_migrate\Plugin\migrate\source\BaseNodeSource;
-use Drupal\sitenow_migrate\Plugin\migrate\source\ProcessMediaTrait;
-
-/**
- * Migrate Source plugin.
- *
- * @MigrateSource(
- *   id = "sppa_potential_employer",
- *   source_module = "node"
- * )
- */
-class PotentialEmployer extends BaseNodeSource {
-  use ProcessMediaTrait;
-
-  /**
-   * {@inheritdoc}
-   */
-  public function query() {
-    $query = parent::query();
-    $query->leftJoin('url_alias', 'alias', "alias.source = CONCAT('node/', n.nid)");
-    $query->fields('alias', ['alias']);
-    return $query;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function fields() {
-    $fields = parent::fields();
-    $fields['alias'] = $this->t('The URL alias for this node.');
-    return $fields;
-  }
-
-}
